renamed `resources` to `schemas`
